// src/lib/lending/types.ts
export interface PoolMetrics {
  marketId: string;
  totalSupplied: number;
  totalBorrowed: number;
  totalCollateralLocked?: number;
  totalShares?: number;
  availableLiquidity: number;
  utilizationRate: number;
  borrowApr: number;
  supplyApr: number;
  supplyApy: number;
  globalYieldIndex: number;
  reserveFactor: number;
  lastIndexUpdate: Date;
}